+++ final_tester/vector_test/test_vector_copy_constructor.cpp
@@ -6,7 +6,7 @@
 /*   By: cerdelen <cerdelen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/29 19:18:34 by cerdelen          #+#    #+#             */
-/*   Updated: 2022/10/27 16:29:46 by cerdelen         ###   ########.fr       */
+/*   Updated: 2022/11/01 10:45:15 by cerdelen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,14 +16,14 @@ void	test_copy_constructor(scope_diff *vars)
 {
 	big_test_header("Testing Copy Constructor");
 
-	int	max_size = 20;
-	std::vector<int>	source;
+	size_t	max_size = 20;
+	std::vector<size_t>	source;
 	source.reserve(max_size);
 	for (size_t i = 0; i < max_size; i++)
 	{
 		source.push_back(i);
 	}
-	ft::vector<int>	source2;
+	ft::vector<size_t>	source2;
 	source2.reserve(max_size);
 	for (size_t i = 0; i < max_size; i++)
 	{
@@ -32,8 +32,8 @@ void	test_copy_constructor(scope_diff *vars)
 	
 	{
 		small_test_header("Standard Copy Constructor");
-		std::vector<int>	og(source);
-		ft::vector<int>		my(source2);
+		std::vector<size_t>	og(source);
+		ft::vector<size_t>		my(source2);
 		
 				for (size_t i = 0; i < max_size; i++)
 		{
@@ -45,11 +45,11 @@ void	test_copy_constructor(scope_diff *vars)
 	}
 
 	{
-		std::vector<int>		og_empty;
-		ft::vector<int>			my_empty;
+		std::vector<size_t>		og_empty;
+		ft::vector<size_t>			my_empty;
 		small_test_header("Empty Copy Constructor");
-		std::vector<int>		og(og_empty);
-		ft::vector<int>			my(my_empty);
+		std::vector<size_t>		og(og_empty);
+		ft::vector<size_t>			my(my_empty);
 
 		check_cap_and_size(&og, &my, "Testing empty copy Constructor", vars);
 		check_scope_diff(vars, "Differences in size or capacity Testing Range Constructor");
